Passes timeout for iocontroller from component
